New compas cluster & level indicator
Added a new compass cluster to show the elevation and temperature. The
day and time have been move as well to make things more readable.

Shortened the exp bar and added a small level indicator. Its small and
stands out very well.

Adjusted the size of the target window name and added a border around
the target health bar for better visibility.